# backend/apps/api_gateway/src/routes/quotations.py
import uuid
from datetime import date
from decimal import Decimal

from fastapi import APIRouter, Depends, HTTPException
from pydantic import BaseModel

from database.session import get_db_session
from modules.core.infrastructure.uow import SQLAlchemyUnitOfWork
from modules.identity.dependencies import require_tenant
from modules.pricing.application.services.pricing_service import PricingService
from modules.pricing.domain.services.price_calculation_engine import PriceCalculationEngine
from modules.pricing.infrastructure.repositories.pricing_repository import PricingRepository
from modules.quotations.application.services.quotation_service import QuotationService
from modules.quotations.infrastructure.adapters.catalog_gateway_impl import CatalogGatewayImpl
from modules.quotations.infrastructure.adapters.pricing_gateway_impl import PricingGatewayImpl
from modules.quotations.infrastructure.repositories.quotation_repository import QuotationRepository

router = APIRouter(prefix="/quotations", tags=["Quotations"])


class CreateQuotationRequest(BaseModel):
    company_id: uuid.UUID
    price_table_id: uuid.UUID
    validity_days: int = 30
    destination_id: uuid.UUID | None = None
    mtr_id: uuid.UUID | None = None
    freight_distance: Decimal | None = None
    freight_cost: Decimal | None = None


class AddQuotationItemRequest(BaseModel):
    service_offering_id: uuid.UUID
    unit_of_measure_id: uuid.UUID
    quantity: Decimal


class CalculateQuotationRequest(BaseModel):
    reference_date: date


def get_quotation_service(session=Depends(get_db_session)) -> QuotationService:
    uow = SQLAlchemyUnitOfWork(session)
    repo = QuotationRepository(session)
    
    # Instantiate pricing pieces for the gateway
    pricing_repo = PricingRepository(session)
    calculation_engine = PriceCalculationEngine()
    
    # Note: PricingService also uses UOW now, but since we are just passing it to the gateway, 
    # we provide the same UOW to PricingService. Wait, let's look at PricingService constructor.
    # We refactored PricingService to take UOW, so we pass it.
    pricing_service = PricingService(uow=uow, repository=pricing_repo, calculation_engine=calculation_engine)
    
    pricing_gateway = PricingGatewayImpl(pricing_service)
    catalog_gateway = CatalogGatewayImpl(session)
    
    return QuotationService(uow, repo, pricing_gateway, catalog_gateway)


@router.post("")
async def create_quotation(
    request: CreateQuotationRequest,
    tenant_id: uuid.UUID = Depends(require_tenant),
    service: QuotationService = Depends(get_quotation_service)
) -> dict:
    quotation_id = await service.create_quotation(
        tenant_id=tenant_id,
        company_id=request.company_id,
        price_table_id=request.price_table_id,
        validity_days=request.validity_days,
        destination_id=request.destination_id,
        mtr_id=request.mtr_id,
        freight_distance=request.freight_distance,
        freight_cost=request.freight_cost
    )
    return {"quotation_id": quotation_id}


from modules.quotations.application.use_cases.get_quotation import GetQuotation
from modules.quotations.application.use_cases.list_quotations import (
    ListQuotations,
    QuotationResponse,
)


@router.get("", response_model=list[QuotationResponse])
async def list_quotations(
    tenant_id: uuid.UUID = Depends(require_tenant),
    session = Depends(get_db_session)
) -> list[QuotationResponse]:
    repo = QuotationRepository(session)
    use_case = ListQuotations(repo)
    return await use_case.execute(tenant_id)


@router.get("/{quotation_id}", response_model=QuotationResponse)
async def get_quotation(
    quotation_id: uuid.UUID,
    tenant_id: uuid.UUID = Depends(require_tenant),
    session = Depends(get_db_session)
) -> QuotationResponse:
    repo = QuotationRepository(session)
    use_case = GetQuotation(repo)
    quotation = await use_case.execute(tenant_id, quotation_id)
    if not quotation:
        raise HTTPException(status_code=404, detail="Quotation not found")
    return quotation


@router.post("/{quotation_id}/items")
async def add_quotation_item(
    quotation_id: uuid.UUID,
    request: AddQuotationItemRequest,
    tenant_id: uuid.UUID = Depends(require_tenant),
    service: QuotationService = Depends(get_quotation_service)
) -> dict:
    try:
        item_id = await service.add_item(
            quotation_id=quotation_id,
            service_offering_id=request.service_offering_id,
            unit_of_measure_id=request.unit_of_measure_id,
            quantity=request.quantity
        )
        return {"item_id": item_id}
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.post("/{quotation_id}/calculate")
async def calculate_quotation(
    quotation_id: uuid.UUID,
    request: CalculateQuotationRequest,
    tenant_id: uuid.UUID = Depends(require_tenant),
    service: QuotationService = Depends(get_quotation_service)
) -> dict:
    try:
        await service.calculate(
            quotation_id=quotation_id,
            reference_date=request.reference_date
        )
        return {"message": "Quotation priced successfully"}
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.post("/{quotation_id}/submit")
async def submit_quotation(
    quotation_id: uuid.UUID,
    tenant_id: uuid.UUID = Depends(require_tenant),
    service: QuotationService = Depends(get_quotation_service)
) -> dict:
    try:
        await service.submit(quotation_id=quotation_id)
        return {"message": "Quotation submitted successfully"}
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.post("/{quotation_id}/approve")
async def approve_quotation(
    quotation_id: uuid.UUID,
    tenant_id: uuid.UUID = Depends(require_tenant),
    service: QuotationService = Depends(get_quotation_service)
) -> dict:
    try:
        await service.approve(quotation_id=quotation_id)
        return {"message": "Quotation approved successfully"}
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.post("/{quotation_id}/reject")
async def reject_quotation(
    quotation_id: uuid.UUID,
    tenant_id: uuid.UUID = Depends(require_tenant),
    service: QuotationService = Depends(get_quotation_service)
) -> dict:
    try:
        await service.reject(quotation_id=quotation_id)
        return {"message": "Quotation rejected successfully"}
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))